requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0,<2.0.0
python-dotenv==0.19.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
requests==2.26.0
aiohttp==3.8.1
twilio==7.11.0
sendgrid==6.9.0
pytest==6.2.5
pytest-asyncio==0.15.1

# AI and ML
openai>=1.0.0
scipy>=1.7.0
tensorflow>=2.7.0
torch>=1.9.0

# Monitoring and Analytics
prometheus-client>=0.11.0
statsd>=3.3.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1

# Utilities
python-jose[cryptography]>=3.3.0
fastapi==0.88.0
pydantic==1.10.8
starlette==0.22.0
uvicorn==0.15.0
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
httpx==0.23.0

# Framework & API
fastapi==0.88.0
uvicorn==0.20.0
pydantic==1.10.2
starlette==0.22.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
pydantic-settings==2.0.3

# Database
sqlalchemy==1.4.42
alembic==1.13.1
psycopg2-binary==2.9.5
asyncpg==0.27.0

# Auth & Security
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0

# File Handling
aiofiles==24.1.0

# HTTP Clients
httpx==0.23.0
requests==2.31.0
aiohttp==3.11.18

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80

# AI & Machine Learning
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
tensorflow-hub==0.14.0
roboflow==1.1.16

# Caching & Performance
redis==5.0.1
orjson==3.9.15

# Utilities
pyjwt==2.10.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
pandas==2.1.4
Markdown==3.5.1

# Testing
pytest==7.3.1
pytest-asyncio==0.21.1

# Add reportlab for PDF generation
reportlab==3.6.12
